- Added several FlatBuffer schema files for new content types. - Reworked task system to use a new fiber-based occupancy and task synchronization system (see Scheduler/Win32/FiberScheduler.hpp and Scheduler/Win32/FiberScheduler.cpp) - Style changes to enums. - Removed unused project files for old MSVC versions.

+++ src/Utility/Concurrency/UserSemaphore.hpp
@@ -20,7 +20,7 @@ namespace Utility {
 
 	class ETPureAbstractHint UserSemaphore : private Utility::Noncopyable {
 	public:
-		virtual size_t	IncreaseCount( const size_t count = 1u ) abstract;
+		virtual int		IncreaseCount( const int count = 1u ) abstract;
 
 		virtual void	Acquire() abstract;
